+++ src/app/about/components/personCard.tsx
