+++ src/app/about/components/personDetailes.tsx
